Add data point and data vector
A data point represents a single point in an n dimensional space.
The dimension represents the number of dimensions the points represents.
This point describes one line in the raw data file

A data vector represents a collection of data points. The points have to be of the
same dimension (determined by the first point). It is possible to extract individual
dimensions from the data vector