Fix error in special point functions
The special point functions

* min_point
* max_point
* center_point

used the wrong value for determining the dimension of the point to create.
Instead of the same dimensions of the vector they used the length of the
vector, which was plain wrong.

With this change the logic of creating the point was extracted into its
own function which itself takes a function to determine how to create create
the point.

The tests were updated to now test for the bug that was overlooked
previously